[log] add compile-time check for `printf` style arg check to log functions
This commit adds compile-time check of format string to all log
functions and macros. It also updates different core modules to make
arg string formats consistent under different platforms. In particular
we use `'%lu` for `uint32_t` arguments and they are explicitly cast
to `unsigned long` using `ToUlong()`.

+++ src/core/net/tcp6_ext.cpp
@@ -170,7 +170,7 @@ void TcpCircularSendBuffer::HandleForwardProgress(size_t aInSendBuffer)
     size_t bytesUntilWrap;
 
     OT_ASSERT(aInSendBuffer <= mCapacityUsed);
-    LogDebg("Forward progress: %u bytes in send buffer\n", aInSendBuffer);
+    LogDebg("Forward progress: %u bytes in send buffer\n", static_cast<unsigned>(aInSendBuffer));
     bytesRemoved   = mCapacityUsed - aInSendBuffer;
     bytesUntilWrap = mCapacity - mStartIndex;
